an admin can't demote or remove herself

+++ controllers/UsersController.php
@@ -32,18 +32,30 @@ class UsersController extends Controller{
                 $db = new DbEntry('UserTable', $this->db);
                 $usr = $db->getRow($this->request['demote_user']);
 
-                //set the new value and save
-                $usr->admin = 0;
-                $usr->save();
+                //admin can't demote herself
+                if ($usr->id == $user->id){
+                    if (!isset($this->data['errors']) || !is_array($this->data['errors'])) $this->data['errors'] = array();
+                    array_push($this->data['errors'], "You can't demote yourself.");
+                } else {
+                    //set the new value and save
+                    $usr->admin = 0;
+                    $usr->save();
+                }
             }
             
             //if delete was pressed
             if (!empty($this->request['delete_user'])){
                 //fetch the user
                 $db = new DbEntry('UserTable', $this->db);
-                //TODO: this should be done in one request: $db->deleteRow(id)
                 $usr = $db->getRow($this->request['delete_user']);
-                $usr->delete();
+                
+                //admin can't remove herself
+                if ($usr->id == $user->id){
+                    if (!isset($this->data['errors']) || !is_array($this->data['errors'])) $this->data['errors'] = array();
+                    array_push($this->data['errors'], "You can't remove yourself.");
+                } else {
+                    $usr->delete();
+                }
             }
             
             //fetch the users from the db
